requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2
django-phonenumber-field==5.1.0
gunicorn==20.1.0
idna==2.10
phonenumbers==8.12.21
Pillow==8.3.1
psycopg2-binary==2.8.6
PyJWT==1.7.1
PyMySQL==1.0.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
twilio==6.56.0
typing-extensions==3.7.4.3
urllib3==1.26.4
